np.int has been depricated and was causing the tests to break

+++ src/cascade_at/dismod/constants.py
@@ -8,7 +8,7 @@ def enum_to_dataframe(enum_name):
     return pd.DataFrame.from_records(
         np.array(
             [(measure, enum_value.value) for (measure, enum_value) in enum_name.__members__.items()],
-            dtype=np.dtype([("name", object), ("value", np.int)]),
+            dtype=np.dtype([("name", object), ("value", int)]),
         )
     )
